First Try at this after my lesson with Michelle.
I tried using classes that Michelle taught me yesterday, but I could not get it to work for the life of me with this file even with using the debugger. I tried solving the problem another way, but that also didn't work. So I must research more into this. It made sense while she was showing me, but it is a lot harder to put into practice! I'll try again, later after work! This is fun though!